Full mutability
This patch essentially reverts the interfaces to the original set:

- `StreamableInterface`
- `MessageInterface`
- `RequestInterface`
- `ResponseInterface`
- `ServerRequestInterface` (originally `IncomingRequestInterface`)

Most properties are now mutable, with the exception of server and file
parameters in the server request interface (as these are considered
non-computable).

Additionally, this patch adds a new property to the request interface:
the base URL. This property is intended to store the scheme and host,
and optionally authentication and port. The "url" is now considered to
ONLY store the path + query string.
